% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Calculator.R
\name{CNV_Simulator}
\alias{CNV_Simulator}
\title{CNV_Simulator Function infers CNV profiles of cells in terms of posterio probability
based on a fitted two-component Gaussian Mixture Model over the log2(CPM/10+1) expression matrix}
\usage{
CNV_Simulator(
  ExprMat,
  minUMIs = 10,
  minCells = 1,
  minGenes = 100,
  repetitions = 8
)
}
\arguments{
\item{ExprMat}{A genes X cells expression matrix of log2(CPM/10+1) scaled scRNA-seq counts}

\item{minUMIs}{Minimal UMI counts for a gene to be kept over all cells (default = 10)}

\item{minCells}{Minimal expressed cell number for a gene to be kept over all cells (default = 1)}

\item{minGenes}{Minimal region size in terms of genes number within the region (default = 100)}

\item{repetitions}{Loop number of fitting to a GMM to obtain an optimal Gaussian mixture model}
}
\value{
a matrix of posterior probabilities as CNV estimates
}
\description{
CNV_Simulator Function infers CNV profiles of cells in terms of posterio probability
based on a fitted two-component Gaussian Mixture Model over the log2(CPM/10+1) expression matrix
}
\examples{
CNV_Simulator(data)

}
