% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterData.R
\name{FilterData}
\alias{FilterData}
\title{FilterData Function loads in data from 10X}
\usage{
FilterData(
  PathToData = ".",
  FeatureColumn = 2,
  UniqueFeatures = TRUE,
  FeatureSuffixTrim = FALSE,
  MinFeatures = 200,
  MinUMIs = 500
)
}
\arguments{
\item{PathToData}{Folder with barcodes.tsv, features.tsv (or genes.tsv), matrix.mtx. Default is the current folder}

\item{FeatureColumn}{choose the column of genes.tsv or features.tsv for feature identifiers; default is 2}

\item{UniqueFeatures}{Logical for unique feature identifiers (default TRUE)}

\item{FeatureSuffixTrim}{Logicalfor removal of trailing "-1" in cell barcodes (default FALSE)}

\item{MinFeatures}{Threshold to remove those cell barcodes with found features fewer than it (default 200)}

\item{MinUMIs}{threshold to remove those cell barcodes with found UMIs fewer than it (default 500)}
}
\value{
a sparse matrix containing the expression data will be returned.
}
\description{
This function allows you to input 10X single-cell RNAseq profile
(loading sparse data matrices provided by 10X genomics),
and remove some unwanted cells.
}
\examples{
# For output from CellRanger < 3.0
FilterData("outs/filtered_gene_bc_matrices_mex/GRCh38/")
# For output from CellRanger >= 3.0
FilterData("outs/filtered_feature_bc_matrix/")

}
