% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_pos}
\alias{gene_pos}
\title{Chromosomal positions of genes.}
\format{
A data frame with over 20000 rows and 5 variables:
}
\source{
\url{http://apr2020.archive.ensembl.org/}
}
\usage{
data(gene_pos)
}
\description{
A dataset containing genomic positons of over 20,000 genes
}
\examples{
data(gene_pos)
}
\references{
(\href{https://www.ncbi.nlm.nih.gov/pubmed/nnnnnnnn}{PubMed})
}
\keyword{-->}
\keyword{23,}
\keyword{24,}
\keyword{25}}
\keyword{ENSG00000274847}}
\keyword{ID,}
\keyword{MAFIP}}
\keyword{MT}
\keyword{X}
\keyword{Y}
\keyword{\describe{}
\keyword{\item{chromosome_name}{chromosome,}
\keyword{\item{end_position}{end}
\keyword{\item{ensembl_gene_id}{Ensembl}
\keyword{\item{hgnc_symbol}{HGNC}
\keyword{\item{start_position}{start}
\keyword{datasets}
\keyword{e.g.}
\keyword{gene}
\keyword{name,}
\keyword{position}}
\keyword{}}
