% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalignantCellsFinder.R
\name{MalignantCellsFinder}
\alias{MalignantCellsFinder}
\title{MalignantCellsFinder Function identifies putative malignant cells based on inferred CNV profiles of cells
using two-component Gaussian Mixture Models fitted to 10X single cells expression profiles}
\usage{
MalignantCellsFinder(PathToData = ".", pvalue = 0.05, p_adj = 1e-06)
}
\arguments{
\item{PathToData}{Folder with barcodes.tsv, features.tsv (or genes.tsv), matrix.mtx}

\item{pvalue}{P value to define outliers relative to CD45-positive cells. Default = 0.05}

\item{p_adj}{P value to define immune cells and malignant cells. Default = 1e-6}
}
\value{
a vector of predicted malignant cells
}
\description{
MalignantCellsFinder Function identifies putative malignant cells based on inferred CNV profiles of cells
using two-component Gaussian Mixture Models fitted to 10X single cells expression profiles
}
\examples{
MalignantCellsFinder("outs/filtered_feature_bc_matrix/")

}
