% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveDyingErythroid.R
\name{RemoveDyingErythroid}
\alias{RemoveDyingErythroid}
\title{RemoveDyingErythroid Function removes dying cells and erythroid cells}
\usage{
RemoveDyingErythroid(
  ExprMat,
  Signature = NULL,
  MinPct.MT = 0.2,
  MinPct.Erythroid = 0.2,
  MinPct.Signature = 0.002
)
}
\arguments{
\item{ExprMat}{Expression matrix of genes X cells}

\item{Signature}{Signature genes to define a cell type for removal (default NULL)}

\item{MinPct.MT}{Minimal percentage of mitochondrial UMIs for a cell to be dying. default is 20\%.}

\item{MinPct.Erythroid}{Minimal percentage of erythroid UMIs for a cell to be an erythrocyte. default is 20\%.}

\item{MinPct.Signature}{Minimal percentage of signature UMIs for a cell to be a specific cell. default is 0.2\%.}
}
\value{
an expression matrix after removing unwanted cells
}
\description{
RemoveDyingErythroid Function removes dying cells and erythroid cells
}
\examples{
RemoveDyingErythroid(data)

}
