% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSquared.R
\name{RSquared}
\alias{RSquared}
\title{RSquared Function calculates an adjusted pseudo R-squared statistic for a Guausian model fitted via nls()}
\source{
\url{https://stats.stackexchange.com/posts/63097/revisions/}
\url{https://stats.stackexchange.com/questions/48703/what-is-the-adjusted-r-squared-formula-in-lm-in-r-and-how-should-it-be-interpret/}
}
\usage{
RSquared(model)
}
\arguments{
\item{model}{a model fitted via nls()}
}
\value{
a pseudo R-squared and an adjusted R-squared are returned.
}
\description{
RSquared Function calculates an adjusted pseudo R-squared statistic for a Guausian model fitted via nls()
}
\examples{
RSquared(model)

}
\references{
Yin, P., & Fan, X. (2001). Estimating R2 shrinkage in multiple regression:
A comparison of different analytical methods. The Journal of Experimental Education, 69(2), 203-224.
}
